Make application running

+++ src/App/Server/GoldEx.Server.Infrastructure/Services/Price/DTOs/TalaIr/TalaIrApiResponseMapper.cs
@@ -29,7 +29,7 @@ public static class TalaIrApiResponseMapper
         {
             foreach (var coin in response.CoinPrice)
             {
-                if (double.TryParse(coin.Value.Value.Replace(",", ""), NumberStyles.Any, CultureInfo.InvariantCulture, out var currentValue))
+                if (decimal.TryParse(coin.Value.Value.Replace(",", ""), NumberStyles.Any, CultureInfo.InvariantCulture, out var currentValue))
                 {
                     priceResponses.Add(new PriceResponse(
                         coin.Value.Title.ToPersianChars(), // Title (Coin Name)
@@ -58,7 +58,7 @@ public static class TalaIrApiResponseMapper
         {
             foreach (var gold in response.GoldPrice)
             {
-                if (double.TryParse(gold.Value.Value.Replace(",", ""), NumberStyles.Any, CultureInfo.InvariantCulture, out var currentValue))
+                if (decimal.TryParse(gold.Value.Value.Replace(",", ""), NumberStyles.Any, CultureInfo.InvariantCulture, out var currentValue))
                 {
                     priceResponses.Add(new PriceResponse(
                         gold.Value.Title.ToPersianChars(), // Title (Gold Name)
@@ -88,7 +88,7 @@ public static class TalaIrApiResponseMapper
         {
             foreach (var currency in response.CurrencyPrice)
             {
-                if (double.TryParse(currency.Value.Value.Replace(",", ""), NumberStyles.Any, CultureInfo.InvariantCulture, out var currentValue))
+                if (decimal.TryParse(currency.Value.Value.Replace(",", ""), NumberStyles.Any, CultureInfo.InvariantCulture, out var currentValue))
                 {
                     priceResponses.Add(new PriceResponse(
                         currency.Value.Title.ToPersianChars(), // Title (Currency Name)
@@ -113,7 +113,7 @@ public static class TalaIrApiResponseMapper
 
         var gold18 = content.GoldPrice.FirstOrDefault(g => g.Key == "gold_18k").Value;
 
-        if (double.TryParse(gold18.Value.Replace(",", ""), NumberStyles.Any, CultureInfo.InvariantCulture, out var currentValue))
+        if (decimal.TryParse(gold18.Value.Replace(",", ""), NumberStyles.Any, CultureInfo.InvariantCulture, out var currentValue))
         {
             return new PriceResponse(
                 gold18.Title.ToPersianChars(), // Title
@@ -136,7 +136,7 @@ public static class TalaIrApiResponseMapper
 
         var dollar = content.CurrencyPrice.FirstOrDefault(c => c.Key == "arz_dolar").Value;
 
-        if (double.TryParse(dollar.Value.Replace(",", ""), NumberStyles.Any, CultureInfo.InvariantCulture, out var currentValue))
+        if (decimal.TryParse(dollar.Value.Replace(",", ""), NumberStyles.Any, CultureInfo.InvariantCulture, out var currentValue))
         {
             return new PriceResponse(
                 dollar.Title.ToPersianChars(), // Title
